Soma small edits
