+++ multibotkit/dispatchers/fb.py
@@ -1,4 +1,4 @@
-from typing import Callable
+from typing import Callable, Optional, Union
 
 from multibotkit.dispatchers.base_dispatcher import BaseDispatcher
 from multibotkit.schemas.fb.incoming import IncomingEvent
@@ -6,11 +6,10 @@ from multibotkit.schemas.fb.incoming import IncomingEvent
 
 class FacebookDispatcher(BaseDispatcher):
 
-    @BaseDispatcher.process_event_decorator
-    async def process_event(
+    def __getting_func_result(
         self,
         event: IncomingEvent,
-        func: Callable = None
+        func: Optional[Callable] = None
     ):
         if func is not None:
             try:
@@ -19,3 +18,11 @@ class FacebookDispatcher(BaseDispatcher):
                 func_result = False
         
         return func_result
+    
+    async def process_event(
+            self,
+            event: IncomingEvent,
+            func: Optional[Callable] = None,
+        ):
+
+        await super().process_event(event, func)
