+++ multibotkit/dispatchers/vk.py
@@ -1,15 +1,14 @@
-from typing import Callable
+from typing import Callable, Optional
 from multibotkit.dispatchers.base_dispatcher import BaseDispatcher
 from multibotkit.schemas.vk.incoming import IncomingEvent
 
 
 class VkontakteDispatcher(BaseDispatcher):
     
-    @BaseDispatcher.process_event_decorator
-    async def process_event(
+    def __getting_func_result(
         self,
         event: IncomingEvent,
-        func: Callable = None
+        func: Optional[Callable] = None
     ):
         if func is not None:
             func_result = False
@@ -20,3 +19,11 @@ class VkontakteDispatcher(BaseDispatcher):
                     func_result = False
         
         return func_result
+    
+    async def process_event(
+            self,
+            event: IncomingEvent,
+            func: Optional[Callable] = None,
+        ):
+
+        await super().process_event(event, func)
